+++ packages/jmespath/tests/unit/compliance/identifiers.test.ts
@@ -1,8 +1,4 @@
-/**
- * Test Compliance with the JMESPath specification
- *
- * @group unit/jmespath/compliance/identifiers
- */
+import { describe, expect, it } from 'vitest';
 import { search } from '../../../src/index.js';
 
 describe('Identifiers tests', () => {