+++ packages/logger/src/types/index.ts
@@ -1,19 +1,13 @@
 export type {
-  LogAttributesWithMessage,
-  LogAttributeValue,
   Environment,
   LogLevelThresholds,
   LogAttributes,
   LogLevel,
-} from './Log.js';
-export type {
   LogItemMessage,
   LogItemExtraInput,
-  LambdaFunctionContext,
-  UnformattedAttributes,
-  PowertoolsLogData,
   ConstructorOptions,
   InjectLambdaContextOptions,
   CustomJsonReplacerFn,
   LoggerInterface,
 } from './Logger.js';
+export type { UnformattedAttributes } from './logKeys.js';