+++ vitest.config.ts
@@ -13,11 +13,11 @@ export default defineConfig({
       include: ['packages/*/src/**'],
       exclude: [
         ...coverageConfigDefaults.exclude,
-        'packages/commons/**',
         'packages/batch/src/types.ts',
+        'packages/commons/src/types/**',
         'packages/event-handler/**',
         'packages/idempotency/**',
-        'packages/jmespath/**',
+        'packages/jmespath/src/types.ts',
         'packages/logger/**',
         'packages/metrics/**',
         'packages/parameters/**',